refacto(all): refacto all the project, creating one build.sbt  combining all projects, cleaning dependencies, data case class and removing logs
