--- alert_consumer/src/main/scala/alert/Event.scala
@@ -1,13 +1,0 @@
-package alert 
-
-import java.time.LocalDateTime
-
-final case class Coords(latitude: Double, longitude: Double)
-
-final case class Event(
-    peacewatcher_id : Int,
-    timestamp: LocalDateTime,
-    location: Coords,
-    words: List[String],
-    persons: List[Person]
-)
