--- alert_consumer/src/main/scala/alert/Person.scala
@@ -1,6 +1,0 @@
-package alert 
-
-final case class Person(
-    name: String,
-    peacescore: Double,
-)
