+++ data/src/main/scala/Alert.scala
@@ -1,6 +1,7 @@
-package alert 
+package data
 
 import java.time.LocalDateTime
+import play.api.libs.json.{Json, OFormat}
 
 final case class Alert(
     peacewatcher_id : Int,
@@ -9,3 +10,7 @@ final case class Alert(
     words: List[String],
     persons: List[Person]
 )
+
+object Alert {
+  implicit val AlertFormatter: OFormat[Alert] = Json.format[Alert]
+}
