+++ data/src/main/scala/Person.scala
@@ -1,0 +1,11 @@
+package data
+
+import play.api.libs.json.{Json, OFormat}
+
+final case class Person(
+                        name: String,
+                        peacescore: Double,
+                       )
+object Person {
+    implicit val PersonFormatter: OFormat[Person] = Json.format[Person]
+}
